+++ src/gen_iter.rs
@@ -5,6 +5,26 @@ use core::pin::Pin;
 
 /// an iterator that holds an internal coroutine representing
 /// the iteration state
+/// 
+/// # Example
+/// pin a self-referenced coroutine in heap, then use it as `Iterator`
+/// ```
+/// #![feature(coroutines)]
+///
+/// use gen_iter::GenIter;
+/// use std::boxed::Box;
+///
+/// let arr = [1, 2];
+/// let c = Box::pin(#[coroutine] static move || {
+///     let arr = &arr;
+///     for i in 0..arr.len() {
+///        yield arr[i];
+///     }
+/// });
+/// let mut g = GenIter(c);
+///
+/// assert_eq!(g.collect::<Vec<i32>>(), [1, 2]);
+/// ```
 #[derive(Copy, Clone, Debug)]
 pub struct GenIter<T>(pub T)
 where
@@ -30,14 +50,15 @@ where
     G: Coroutine<Return = ()> + Unpin,
 {
     #[inline]
-    fn from(gen: G) -> Self {
-        GenIter(gen)
+    fn from(g: G) -> Self {
+        GenIter(g)
     }
 }
 
 
 /// macro to simplify iterator - via - coroutine construction
 ///
+/// - create a movable coroutine as `Iterator`
 /// ```
 /// #![feature(coroutines)]
 ///
@@ -48,10 +69,24 @@ where
 ///     yield 2;
 /// });
 ///
-/// assert_eq!(g.next(), Some(1));
-/// assert_eq!(g.next(), Some(2));
-/// assert_eq!(g.next(), None);
+/// assert_eq!(g.collect::<Vec<i32>>(), [1, 2]);
+/// ```
+/// 
+/// - create an immovable coroutine (self-referenced) pinned in stack as `Iterator`
+/// ```
+/// #![feature(coroutines)]
 ///
+/// use gen_iter::gen_iter;
+///
+/// let arr = [1, 2];
+/// let mut g = gen_iter!(static move {
+///     let arr = &arr;
+///     for i in 0..arr.len() {
+///        yield arr[i];
+///     }
+/// });
+///
+/// assert_eq!(g.collect::<Vec<i32>>(), [1, 2]);
 /// ```
 #[macro_export]
 macro_rules! gen_iter {
@@ -60,7 +95,14 @@ macro_rules! gen_iter {
     };
     (move $block: block) => {
         $crate::GenIter(#[coroutine] move || $block)
-    }
+    };
+
+    (static $block: block) => {
+        $crate::GenIter { 0: ::core::pin::pin!(#[coroutine] static || $block) }
+    };
+    (static move $block: block) => {
+        $crate::GenIter { 0: ::core::pin::pin!(#[coroutine] static move || $block) }
+    };
 }
 
 
@@ -102,6 +144,52 @@ mod tests {
             yield 2;
         });
 
+        assert_eq!(g.next(), Some(1));
+        assert_eq!(g.next(), Some(2));
+        assert_eq!(g.next(), None);
+    }
+
+    #[test]
+    fn self_ref_coroutine_in_stack() {
+        let c = ::core::pin::pin!(#[coroutine] static || {
+            let v1 = [1, 2];
+            let v = &v1;
+            for i in 0..v.len() {
+                yield v[i];
+            }
+        });
+        let mut g = GenIter(c);
+        
+        assert_eq!(g.next(), Some(1));
+        assert_eq!(g.next(), Some(2));
+        assert_eq!(g.next(), None);
+    }
+
+    #[test]
+    fn gen_iter_macro_static() {
+        let mut g = gen_iter!(static {
+            let v1 = [1, 2];
+            let v = &v1;
+            for i in 0..v.len() {
+                yield v[i];
+            }
+        });
+        
+        assert_eq!(g.next(), Some(1));
+        assert_eq!(g.next(), Some(2));
+        assert_eq!(g.next(), None);
+    }
+
+    #[test]
+    fn gen_iter_macro_static_move() {
+        let v1 = [1, 2];
+        let mut g = gen_iter!(static move {
+            let v = &v1;
+            for i in 0..v.len() {
+                yield v[i];
+            }
+        });
+        
         assert_eq!(g.next(), Some(1));
         assert_eq!(g.next(), Some(2));
         assert_eq!(g.next(), None);